fix second example
